backend/requirements.txt
asgiref==3.11.1
Django==4.2.30
django-cors-headers==4.9.0
djangorestframework==3.16.1
django-filter==24.3
Faker==40.15.0
mysqlclient==2.2.7
python-dotenv==1.2.2
sqlparse==0.5.5
toposort==1.10
typing_extensions==4.15.0
